Use multiple lines (not tested yet...)

+++ src/main/scala/parser/ratp/Transfer.scala
@@ -1,12 +1,12 @@
 package parser.ratp
 
-case class Transfer(fromStopId: Long, toStopId: Long, transferType: String, minTransferTime: Int)
+case class Transfer(fromStopId: Int, toStopId: Int, transferType: String, minTransferTime: Int)
 
 object Transfer {
   def parse(fields: Map[String, String]) = {
     Transfer(
-      fields("from_stop_id").toLong,
-      fields("to_stop_id").toLong,
+      fields("from_stop_id").toInt,
+      fields("to_stop_id").toInt,
       fields("transfer_type"),
       fields("min_transfer_time").toInt
     )
